+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+### Fix
+
+- Update authors so that we appear as the contact on PyPi instead of Bede.
+- PyPi URL now points to new Pathogena project.
+
 ## 2.0.0rc1 (2024-07-18)
 
 ### Feat